# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "StockTank API",
    "version": "0.1.0",
    "description": "Identity, RBAC, public content, advertising, marketing and system endpoints for StockTank. Sessions are cookie-based; state-changing requests must send `X-Requested-With`. Financial content served by this platform is informational only and never investment advice."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin (dev proxy or deployed API host)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "st_session",
        "description": "httpOnly session cookie set by /api/v1/auth/register and /api/v1/auth/login. The server stores only a SHA-256 hash of the token."
      }
    },
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "enum": [
                  "BAD_REQUEST",
                  "VALIDATION_FAILED",
                  "UNAUTHENTICATED",
                  "FORBIDDEN",
                  "NOT_FOUND",
                  "CONFLICT",
                  "RATE_LIMITED",
                  "INTERNAL"
                ]
              },
              "message": {
                "type": "string"
              },
              "requestId": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ],
        "description": "Envelope used by every non-2xx response.",
        "example": {
          "error": {
            "code": "VALIDATION_FAILED",
            "message": "Invalid request body",
            "requestId": "7f1c1b7e-2d1a-4a4d-9d2b-3b7b2a1f0c9e",
            "details": {
              "source": "body",
              "issues": [
                {
                  "path": "email",
                  "message": "Invalid email address",
                  "code": "invalid_format"
                }
              ]
            }
          }
        }
      },
...
      "NewsletterSubscribeRequest": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "maxLength": 254,
            "format": "email"
          },
          "consent": {
            "type": "boolean",
            "enum": [
              true
            ]
          },
          "source": {
            "type": "string",
            "maxLength": 60
          },
          "utm": {
            "type": "object",
            "properties": {
              "source": {
                "type": "string",
                "maxLength": 100
              },
              "medium": {
                "type": "string",
                "maxLength": 100
              },
              "campaign": {
                "type": "string",
                "maxLength": 100
              }
            }
          },
          "website": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "email",
          "consent"
        ],
        "example": {
          "email": "reader@example.com",
          "consent": true,
          "source": "homepage"
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/v1/newsletter/subscribe": {
      "post": {
        "tags": [
          "Marketing"
        ],
        "summary": "Subscribe to the newsletter",
        "description": "Double opt-in: emails a confirmation link. The response is identical for new and existing addresses.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
              "example": "stocktank"
            },
            "required": true,
            "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
            "name": "x-requested-with",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewsletterSubscribeRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Check your inbox",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "check_inbox"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "No email provider configured or delivery failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/newsletter/confirm": {
      "post": {
        "tags": [
          "Marketing"
        ],
        "summary": "Confirm a subscription",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
              "example": "stocktank"
            },
            "required": true,
            "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
            "name": "x-requested-with",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 200
                  }
                },
                "required": [
                  "token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Done",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "confirmed",
                        "unsubscribed"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Token not found (NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/newsletter/unsubscribe": {
      "post": {
        "tags": [
          "Marketing"
        ],
        "summary": "Unsubscribe",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
              "example": "stocktank"
            },
            "required": true,
            "description": "CSRF guard required on every state-changing request under /api. Any non-empty value (the official clients send \"stocktank\").",
            "name": "x-requested-with",
            "in": "header"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": {
                    "type": "string",
                    "minLength": 20,
                    "maxLength": 200
                  }
                },
                "required": [
                  "token"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Done",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "enum": [
                        "confirmed",
                        "unsubscribed"
                      ]
                    }
                  },
                  "required": [
                    "status"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Malformed or invalid input (BAD_REQUEST / VALIDATION_FAILED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Token not found (NOT_FOUND)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded (RATE_LIMITED)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error (INTERNAL)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}